+++ csharp/bypstest-ser/src-ser/com/wilutions/byps/test/api/BSerializer_1516687588.cs
@@ -28,7 +28,7 @@ namespace com.wilutions.byps.test.api
 			return bmodel.computeArrayWireSize(elementSize, n);
 		}
 		
-		public override Object read(Object obj1, BInput bin1, int version) {
+		public override Object read(Object obj1, BInput bin1, long version) {
 			BInputBin bin = (BInputBin)bin1;
 			BBufferBin bbuf = bin.bbuf;
 			
@@ -55,7 +55,7 @@ namespace com.wilutions.byps.test.api
 			return arr;
 		}
 		
-		public override void write(Object obj1, BOutput bout1, int version) {
+		public override void write(Object obj1, BOutput bout1, long version) {
 			BOutputBin bout = (BOutputBin)bout1;
 			BBufferBin bbuf = bout.bbuf;
