+++ csharp/bypstest-ser/src-ser/com/wilutions/byps/test/api/BSerializer_94341197.cs
@@ -18,7 +18,7 @@ namespace com.wilutions.byps.test.api
 		public BSerializer_94341197()
 			: base(94341197) {}		
 		
-		public override Object read(object obj1, BInput bin1, int version) {
+		public override Object read(object obj1, BInput bin1, long version) {
 			BInputBin bin = ((BInputBin)bin1);
 			BBufferBin bbuf = bin.bbuf;
 			Dictionary<char,float> map = new Dictionary<char,float>();
@@ -32,7 +32,7 @@ namespace com.wilutions.byps.test.api
 			return map;
 		}
 		
-		public override void write(Object obj1, BOutput bout1, int version)  {
+		public override void write(Object obj1, BOutput bout1, long version)  {
 			BOutputBin bout = ((BOutputBin)bout1);
 			BBufferBin bbuf = bout.bbuf;
 			Dictionary<char,float> map = (Dictionary<char,float>)obj1;
