+++ csharp/bypstest-ser/src-ser/com/wilutions/byps/test/api/inherit/BSerializer_6001.cs
@@ -31,7 +31,7 @@ namespace com.wilutions.byps.test.api.inherit
 			       /* padding up to multiple of alignment */ + 0;			
 		}
 		
-		public override void write(Object obj1, BOutput bout1, int version)
+		public override void write(Object obj1, BOutput bout1, long version)
 		{
 			Class1 obj = (Class1)obj1;			
 			BOutputBin bout = (BOutputBin)bout1;
@@ -40,7 +40,7 @@ namespace com.wilutions.byps.test.api.inherit
 			bout.writeObj(obj.NextClass1, false, null);
 		}
 		
-		public override Object read(Object obj1, BInput bin1, int version)
+		public override Object read(Object obj1, BInput bin1, long version)
 		{
 			BInputBin bin = (BInputBin)bin1;
 			Class1 obj = (Class1)(obj1 != null ? obj1 : bin.onObjectCreated(new Class1()));
