+++ csharp/bypstest-ser/src-ser/com/wilutions/byps/test/api/remote/BSerializer_1000850298.cs
@@ -29,7 +29,7 @@ namespace com.wilutions.byps.test.api.remote
 			       /* padding up to multiple of alignment */ + 0;			
 		}
 		
-		public override Object read(Object obj1, BInput bin1, int version)
+		public override Object read(Object obj1, BInput bin1, long version)
 		{
 			BInputBin bin = (BInputBin)bin1;
 			BRequest_RemoteSetTypes_getShort1 obj = (BRequest_RemoteSetTypes_getShort1)(obj1 != null ? obj1 : bin.onObjectCreated(new BRequest_RemoteSetTypes_getShort1()));
