+++ java/bypstest-ser/src-ser-bin/com/wilutions/byps/test/api/BSerializer_309901669.java
@@ -36,7 +36,7 @@ public class BSerializer_309901669 extends BSerializer {
 	
 	
 	@Override
-	public Object read(final Object obj1, final BInput bin1, final int version) throws BException {
+	public Object read(final Object obj1, final BInput bin1, final long version) throws BException {
 		final BInputBin bin = (BInputBin)bin1;
 		final BResult_19 obj = (BResult_19)(obj1 != null ? obj1 : bin.onObjectCreated(new BResult_19()));
